Fix soft 404 detection
SPAs serve a 200 with HTML instead of a 404 for unknown URLs (they really shouldn't do that for static URLs, but that's another story.)
We tried to detect that, but the regexp was buggy: It required a space at the start.
Instead of looking at the HTTP body, look at the content type.