Fix timeout handling
From my home connection, I'm seeing infrequent timeouts.
As far as I can see, this is an artefact of the chunk-tester sending hundreds of requests in the same millisecond, and the connection not being fast enough to get them all through within 10s (bump timeout to 20s to avoid that).

Instead, retry twice upon a timeout.

Log to console so that we can later see how often that stuff happens, but do not clutter the client display.